divide into usecase
