+++ MutaroModule/Sources/Features/MyApps/Model/AppInfo.swift
@@ -1,0 +1,14 @@
+//
+//  AppInfo.swift
+//
+//
+//  Created by minguk-kim on 2023/06/04.
+//
+
+import Foundation
+
+public struct AppInfo: Equatable {
+    let id: String
+    let name: String
+    let iconUrl: String?
+}
