--- MutaroModule/Sources/Features/MyApps/Model/MyAppsModel.swift
@@ -1,12 +1,0 @@
-//
-//  MyAppsModel.swift
-//
-//
-//  Created by minguk-kim on 2023/01/04.
-//
-
-import Foundation
-
-public struct MyAppsModel {}
-
-extension MyAppsModel: Equatable {}
