+++ MutaroModule/Sources/Features/MyApps/ViewModel/MyAppsViewModel.swift
@@ -5,7 +5,6 @@
 //  Created by minguk-kim on 2023/01/01.
 //
 
-import Client
 import Combine
 import Core
 import Foundation
@@ -30,110 +29,22 @@ public final class MyAppsViewModel: NSObject, MyAppsViewModelProtocol {
         self.environment = environment
     }
 
-    func loadStoredJWTInfo() {
+    func loadStoredJWTInfo() async {
         do {
             let storedJWTInfo: MutaroJWT.JWTRequestInfo = try KeychainStore.shared.loadValue(forKey: .jwt)
 
-            guard let currentJWTInfo, currentJWTInfo != storedJWTInfo else {
-                shouldShowRegisterJWTSubject.send(false)
-                currentJWTInfo = storedJWTInfo
-                fetchMyApps(storedJWTInfo: storedJWTInfo)
+            if let currentJWTInfo, currentJWTInfo != storedJWTInfo {
+                shouldShowRegisterJWTSubject.send(true)
                 return
             }
 
-            shouldShowRegisterJWTSubject.send(true)
+            shouldShowRegisterJWTSubject.send(false)
+            currentJWTInfo = storedJWTInfo
+            let appInfos = try await environment.appInfoUseCase.fetchAppInfos(storedJWTInfo: storedJWTInfo)
+            appInfosSubject.send(appInfos)
         } catch {
             shouldShowRegisterJWTSubject.send(true)
         }
-    }
-
-    func fetchMyApps(storedJWTInfo: MutaroJWT.JWTRequestInfo) {
-        Task {
-            let builder = MutaroJWT.AppstoreConnectJWTBuilder(
-                keyId: storedJWTInfo.keyID,
-                issuerId: storedJWTInfo.issuerID,
-                pemString: storedJWTInfo.privateKey
-            )
-            guard let token = try? builder.generateJWT() else {
-                // TODO: - show need to register
-                print("Mins: failed generate token")
-                return
-            }
-
-            do {
-                let myApps = try await getMyApps(token: token)
-                let appInfos = try await getAppInfos(token: token, myApps: myApps)
-                appInfosSubject.send(appInfos)
-            } catch {
-                print("Mins: \(error)")
-            }
-        }
-    }
-
-    private func getMyApps(token: String) async throws -> [(String, String)] {
-        let myAppsParameters = [
-            "fields[apps]": "name"
-        ]
-        let myAppsEndpoint = MyAppsEndpoint.GetAllListMyApps(token: token, additionalParameters: myAppsParameters)
-        let myAppsResult = await environment.client.request(
-            endpoint: myAppsEndpoint,
-            responseModel: MyAppsDTO.self
-        )
-        let myAppsResultEntity = try myAppsResult.get().toEntity()
-        let appInfos = myAppsResultEntity.data?
-            .compactMap { data -> (String, String)? in
-                guard let id = data.id,
-                      let name = data.attributes?.name else {
-                    return nil
-                }
-                return (id, name)
-            }
-
-        return appInfos ?? []
-    }
-
-    private func getAppInfos(token: String, myApps: [(String, String)]) async throws -> [AppInfo] {
-        try await myApps
-            .concurrentMap { [weak self] app -> AppInfo? in
-                guard let self else {
-                    return nil
-                }
-                let appId = app.0
-                let appName = app.1
-
-                let buildsParametr: [String: Any] = [
-                    "filter[app]": appId,
-                    "sort": "-uploadedDate",
-                    "limit": 1
-                ]
-                let buildsEndpoint = BuildsEndpoint.GetAllBuilds(token: token, additionalParameters: buildsParametr)
-                let buildsResult = await self.environment.client.request(endpoint: buildsEndpoint, responseModel: BuildsDTO.self)
-                guard let buildsResultEntity = try? buildsResult.get().toEntity(),
-                      let data = buildsResultEntity.data?.first else {
-                    return .init(
-                        id: appId,
-                        name: appName,
-                        iconUrl: nil
-                    )
-                }
-                let iconAsset = data.attributes?.iconAssetToken
-                let width = iconAsset?.width ?? 167
-                let height = iconAsset?.height ?? 167
-                let imageUrl = iconAsset?.templateURL
-                guard let parsedImageUrl = AppleImageTemplateUrlParser.parse(
-                    url: imageUrl,
-                    width: width,
-                    height: height
-                ) else {
-                    return nil
-                }
-                return .init(
-                    id: appId,
-                    name: appName,
-                    iconUrl: parsedImageUrl
-                )
-            }
-            .compactMap { $0 }
     }
 
     func onTapRegisterJWT(from viewController: UIViewController) {
@@ -167,13 +78,5 @@ public final class MyAppsViewModel: NSObject, MyAppsViewModelProtocol {
             }
             environment.imageDownloadService.cancelDownloadImage(with: url)
         }
-    }
-}
-
-extension MyAppsViewModel {
-    struct AppInfo: Equatable {
-        let id: String
-        let name: String
-        let iconUrl: String?
     }
 }
